+++ src/pages/ProductPage.js
@@ -1,30 +1,24 @@
-import {Container, Row, Col} from 'react-bootstrap'
-import 'bootstrap/dist/css/bootstrap.min.css';
-import DetailsProduct from '../components/DetailsProduct/DetailsProduct.js'
-import CartBadge from '../components/CartBadge/CartBadge'
-import useGet from '../hooks/useGet';
-import { URL_PRODUCTS } from '../constants/index.js';
+import DetailsProduct from "../components/DetailsProducts/DetailsProducts";
+import { Container, Row } from "react-bootstrap";
+import { useEffect, useState } from "react";
+import axiosClient from '../config/axiosClient'
 
 const ProductPage = () => {
-    const products = useGet(URL_PRODUCTS);
+  const [product, setProduct] = useState ('');
 
-    return (
-        <Container className='m-5'>
-            <Row>
-                <Col xs={6}>
-                    <img className='w-100' src='https://www.cocinayvino.com/wp-content/uploads/2018/08/pizza-napolitana-2-e1534286138178-1200x675.jpg'/>
-                </Col>
-                <Col xs={6}>
-                    <DetailsProduct
-                    id= {products.id}
-                    title= {products.name}
-                    price = {products.price}
-                    ></DetailsProduct>
-                </Col>
-            </Row>
-                <CartBadge></CartBadge>
-        </Container>
-    );
-}
- 
+  
+
+  useEffect(() => {
+    getProduct();
+  }, []);
+
+  return (
+    <Container className="m-5">
+      <Row>
+        <DetailsProduct></DetailsProduct>
+      </Row>
+    </Container>
+  );
+};
+
 export default ProductPage;
